backend_recipe/requirements.txt
# Core API Framework
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.12.2
pydantic-settings==2.11.0

# Database & Storage
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
supabase==2.22.1
boto3==1.35.94
botocore==1.35.94

# LangChain Ecosystem
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-google-vertexai==3.0.0
langchain-postgres==0.0.16
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37

# Google AI/Cloud
google-genai==1.45.0
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-aiplatform==1.121.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0

# Vector Search & Embeddings
faiss-cpu==1.12.0
pgvector==0.3.6

# PDF Processing
pypdf==6.1.1

# HTTP & Networking
httpx==0.28.1
httpx-sse==0.4.3
requests==2.32.5
aiohttp==3.13.0

# Data Processing
numpy==2.3.4
pandas==2.2.3
pillow==12.0.0

# Utilities
tenacity==9.1.2
pyyaml==6.0.3
click==8.3.0
rich==14.2.0
tqdm==4.67.1

# Security & Auth
cryptography==46.0.3
PyJWT==2.10.1

# Production Monitoring
sentry-sdk==2.42.0
